Corrigido erro #5
Alterado o formato de u16 para u32, permitindo que seja lido códigos de imóvel.
Também adaptado o código para essa diferença.
