+++ src/main.rs
@@ -1,14 +1,26 @@
-mod leitortransf;
-
 use chrono::NaiveDate;
 use clipboard::{ClipboardContext, ClipboardProvider};
 use color_print::{cprint, cprintln};
 use crossterm::terminal::{disable_raw_mode, enable_raw_mode, Clear, ClearType};
 use crossterm::{cursor, execute};
-use leitortransf::ler_código_início;
 use mod_cod_barras::*;
 use rustyline::{DefaultEditor, Result};
 use std::io::stdout;
+
+pub fn ler_código_início(ctx: &mut ClipboardContext) -> String {
+    loop {
+        disable_raw_mode().unwrap();
+        execute!(stdout(), Clear(ClearType::All), cursor::MoveTo(0, 0)).unwrap();
+        cprintln!("<green, bold>Parece que você possui um código de barras copiado, deseja inserir esse?</>");
+        cprintln!("<green, bold><u>S</>im</>/<red, bold><u>N</>ão</>");
+        enable_raw_mode().unwrap();
+        break match read_char().unwrap() {
+            's' | ' ' => String::from(&ctx.get_contents().unwrap()),
+            'n' => String::from(""),
+            _ => continue,
+        };
+    }
+}
 
 fn main() -> Result<()> {
     // rl é um atalho para read_line de rustyline
@@ -144,11 +156,11 @@ fn main() -> Result<()> {
                     break;
                 }
             }
-            'g' => {
+            'g' | 'i' => {
                 disable_raw_mode().unwrap();
 
                 loop {
-                    cprintln!("<green, bold>Digite o novo número da guia:</>");
+                    cprintln!("<green, bold>Digite o novo número da guia/imóvel:</>");
 
                     let Ok(valor_tmp) = rl
                         .readline_with_initial(
